--- src/main/java/net/spb/spb/controller/AdminController.java
@@ -1,910 +1,0 @@
-package net.spb.spb.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import net.spb.spb.dto.ChapterDTO;
-import net.spb.spb.dto.LectureDTO;
-import net.spb.spb.dto.OrderDTO;
-import net.spb.spb.dto.TeacherDTO;
-import net.spb.spb.dto.member.MemberDTO;
-import net.spb.spb.dto.pagingsearch.*;
-import net.spb.spb.dto.post.PostDTO;
-import net.spb.spb.dto.post.PostReportDTO;
-import net.spb.spb.dto.qna.QnaDTO;
-import net.spb.spb.service.AdminService;
-import net.spb.spb.service.ReportService;
-import net.spb.spb.service.board.BoardServiceIf;
-import net.spb.spb.service.qna.QnaService;
-import net.spb.spb.service.teacher.TeacherServiceIf;
-import net.spb.spb.service.member.MemberServiceIf;
-import net.spb.spb.util.*;
-import org.apache.poi.ss.usermodel.Row;
-import org.apache.poi.ss.usermodel.Sheet;
-import org.apache.poi.ss.usermodel.Workbook;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-
-import java.io.File;
-import java.io.IOException;
-import java.util.*;
-import java.util.stream.Collectors;
-
-@Controller
-@Log4j2
-@RequiredArgsConstructor
-@RequestMapping("/admin")
-public class AdminController {
-
-    private final MemberServiceIf memberService;
-    private final ReportService reportService;
-    private final AdminService adminService;
-    private final TeacherServiceIf teacherService;
-    private final BoardServiceIf boardService;
-    private final QnaService qnaService;
-    private final FileUtil fileUtil;
-
-    private static final long MAX_FILE_SIZE_500 = 500 * 1024 * 1024;
-    private static final long MAX_FILE_SIZE_10 = 10 * 1024 * 1024;
-    private static final long MAX_FILE_SIZE_100 = 100 * 1024 * 1024;
-    private static final Map<String, String> ROOT_BREADCRUMB = Map.of("name", "관리 페이지", "url", "/admin");
-
-    @GetMapping({"","/"})
-    public String adminMain(Model model) {
-        setBreadcrumb(model, Map.of("관리 페이지", ""));
-        return "admin/main";
-    }
-
-    @GetMapping("/member/list")
-    public void memberList(@ModelAttribute MemberPageDTO memberPageDTO, Model model, HttpServletRequest req) {
-        String baseUrl = req.getRequestURI();
-        memberPageDTO.setLinkUrl(PagingUtil.buildLinkUrl(baseUrl, memberPageDTO));
-        if (memberPageDTO.getSearch_member_grade() != null && memberPageDTO.getSearch_member_grade().trim().equals("")) {
-            memberPageDTO.setSearch_member_grade(null);
-        }
-        int total_count = memberService.getMemberCount(memberPageDTO);
-        memberPageDTO.setTotal_count(total_count);
-        String paging = NewPagingUtil.pagingArea(memberPageDTO);
-        List<MemberDTO> memberDTOs = memberService.getMembers(memberPageDTO);
-        model.addAttribute("searchDTO", memberPageDTO);
-        model.addAttribute("list", memberDTOs);
-        model.addAttribute("memberTotalCount", total_count);
-        model.addAttribute("paging", paging);
-        setBreadcrumb(model, Map.of("회원 목록", ""));
-    }
-
-    @PostMapping("/member/update")
-    public String memberModify(@ModelAttribute MemberDTO memberDTO) {
-        memberService.updateMemberByAdmin(memberDTO);
-        return "/admin/member/update-success";
-    }
-
-    @PostMapping("/member/state")
-    public String memberModifyState(@ModelAttribute MemberPageDTO memberPageDTO, @ModelAttribute MemberDTO memberDTO) {
-        memberService.updateMemberState(memberDTO);
-        String query = memberPageDTO.toQueryString();
-        return "redirect:/admin/member/list?" + query;
-    }
-
-    @GetMapping("/member/view")
-    public String memberView(@RequestParam("memberId") String memberId, Model model) {
-        MemberDTO memberDTO = memberService.getMemberById(memberId);
-        model.addAttribute("memberDTO", memberDTO);
-        return "/admin/member/memberPopup";
-    }
-
-    @GetMapping("/report/list")
-    public String reportList() {
-        return "admin/report/list";
-    }
-
-    @GetMapping("/report/list/board")
-    public String boardReportList(@ModelAttribute SearchDTO searchDTO,
-                                  @ModelAttribute PageRequestDTO pageRequestDTO,
-                                  Model model) {
-        int totalCount = reportService.boardReportTotalCount(searchDTO);
-        PageResponseDTO<PostReportDTO> pageResponseDTO = PageUtil.buildAndCorrectPageResponse(
-                pageRequestDTO,
-                totalCount,
-                () -> reportService.listBoardReport(searchDTO, pageRequestDTO)
-        );
-
-        model.addAttribute("responseDTO", pageResponseDTO);
-        model.addAttribute("boardReportList", pageResponseDTO.getDtoList());
-        model.addAttribute("searchDTO", searchDTO);
-        setBreadcrumb(model, Map.of("자유게시판 신고 목록", ""));
-
-        return "admin/report/boardReportTable";
-    }
-
-    @PostMapping(path = "/report/board/reject", produces = "text/plain;charset=UTF-8")
-    public ResponseEntity<String> rejectBoardReport(@RequestParam("reportIdx") String reportIdx) {
-        reportService.rejectBoardReport(reportIdx);
-        return ResponseEntity.ok("신고가 반려되었습니다.");
-    }
-
-    @PostMapping(path = "/report/board/process", produces = "text/plain;charset=UTF-8")
-    public ResponseEntity<String> processBoardReport(@RequestParam("reportIdx") String reportIdx) {
-        reportService.updateBoardReportAsProcessed(reportIdx);
-        return ResponseEntity.ok("신고가 처리되었습니다.");
-    }
-
-    @GetMapping("/report/list/review")
-    public String reviewReportList(@ModelAttribute SearchDTO searchDTO,
-                                   @ModelAttribute PageRequestDTO pageRequestDTO,
-                                   Model model) {
-        int totalCount = reportService.reviewReportTotalCount(searchDTO);
-        PageResponseDTO<PostReportDTO> pageResponseDTO = PageUtil.buildAndCorrectPageResponse(
-                pageRequestDTO,
-                totalCount,
-                () -> reportService.listReviewReport(searchDTO, pageRequestDTO)
-        );
-
-        model.addAttribute("responseDTO", pageResponseDTO);
-        model.addAttribute("reviewReportList", pageResponseDTO.getDtoList());
-        model.addAttribute("searchDTO", searchDTO);
-        setBreadcrumb(model, Map.of("강의평 신고 목록", ""));
-
-        return "admin/report/reviewReportTable";
-    }
-
-    @PostMapping(path = "/report/review/process", produces = "text/plain;charset=UTF-8")
-    @ResponseBody
-    public ResponseEntity<String> processReviewReport(@RequestParam("reportIdx") String reportIdx) {
-        reportService.updateReviewReportAsProcessed(reportIdx);
-        return ResponseEntity.ok("신고가 처리되었습니다.");
-    }
-
-    @PostMapping(path = "/report/review/reject", produces = "text/plain;charset=UTF-8")
-    @ResponseBody
-    public ResponseEntity<String> rejectReviewReport(@RequestParam("reportIdx") String reportIdx) {
-        reportService.rejectReviewReport(reportIdx);
-        return ResponseEntity.ok("신고가 반려되었습니다.");
-    }
-
-    @GetMapping("/teacher/list")
-    public void teacherList(@ModelAttribute MemberPageDTO memberPageDTO, Model model, HttpServletRequest req) {
-        List<MemberDTO> teacherWithTeacherProfile = adminService.selectTeacherWithTeacherProfile();
-        List<MemberDTO> teacherWithoutTeacherProfile = adminService.selectTeacherWithoutTeacherProfile();
-        model.addAttribute("teacher1", teacherWithTeacherProfile);
-        model.addAttribute("teacher2", teacherWithoutTeacherProfile);
-        setBreadcrumb(model, Map.of("선생님 목록", ""));
-    }
-
-    @GetMapping("/teacher/regist")
-    public String teacherRegist(@RequestParam(name = "memberId", defaultValue = "") String memberId,
-                                Model model,
-                                RedirectAttributes redirectAttributes,
-                                HttpSession session) {
-
-        if (memberId.isBlank()) {
-            redirectAttributes.addFlashAttribute("errorMessage", "회원 ID가 누락되었습니다.");
-            return "redirect:/admin/teacher/list";
-        }
-
-        MemberDTO memberDTO = memberService.getMemberById(memberId);
-        if (memberDTO == null) {
-            redirectAttributes.addFlashAttribute("errorMessage", "존재하지 않는 회원입니다.");
-            return "redirect:/admin/teacher/list";
-        }
-
-        if(adminService.existsByTeacherId(memberDTO.getMemberId())) {
-            redirectAttributes.addFlashAttribute("errorMessage", "이미 등록된 선생님입니다.");
-            return "redirect:/admin/teacher/list";
-        }
-
-        session.setAttribute("registTeacherName", memberDTO.getMemberName());
-        session.setAttribute("registTeacherId", memberDTO.getMemberId());
-
-        model.addAttribute("memberDTO", memberDTO);
-        setBreadcrumb(model,
-                Map.of("선생님 목록", "/admin/teacher/list"),
-                Map.of("선생님 등록", "")
-        );
-        return "admin/teacher/regist";
-    }
-
-    @PostMapping("/teacher/regist")
-    public String teacherRegistPost(@RequestParam(name = "file1") MultipartFile file,
-                                    @Valid @ModelAttribute TeacherDTO teacherDTO,
-                                    BindingResult bindingResult,
-                                    RedirectAttributes redirectAttributes,
-                                    HttpSession session) {
-
-        // 유효성 검사 시작
-        String sessionTeacherId = (String)session.getAttribute("registTeacherId");
-        String sessionTeacherName = (String)session.getAttribute("registTeacherName");
-
-        if (bindingResult.hasErrors()) {
-            List<String> errorMessages = bindingResult.getFieldErrors().stream()
-                    .map(FieldError::getDefaultMessage)
-                    .collect(Collectors.toList());
-            redirectAttributes.addFlashAttribute("errorMessages", errorMessages);
-            redirectAttributes.addFlashAttribute("teacherDTO", teacherDTO);
-            return "redirect:/admin/teacher/regist?memberId="+sessionTeacherId;
-        }
-
-        if (!sessionTeacherId.equalsIgnoreCase(teacherDTO.getTeacherId()) || !sessionTeacherName.equalsIgnoreCase(teacherDTO.getTeacherName())) {
-            redirectAttributes.addFlashAttribute("errorMessages", "선생님 등록 중에 오류가 발생했습니다.");
-            redirectAttributes.addFlashAttribute("teacherDTO", teacherDTO);
-            return "redirect:/admin/teacher/regist?memberId="+sessionTeacherId;
-        }
-
-        if (file != null && !file.isEmpty() && file.getSize() > (10 * 1024 * 1024)) {
-            redirectAttributes.addFlashAttribute("errorMessage", "파일은 10MB 이하만 업로드할 수 있습니다.");
-            redirectAttributes.addFlashAttribute("teacherDTO", teacherDTO);
-            return "redirect:/admin/teacher/regist?memberId="+sessionTeacherId;
-        }
-        // 유효성 종료
-
-        teacherDTO.setTeacherId(sessionTeacherId);
-        teacherDTO.setTeacherName(sessionTeacherName);
-
-        try {
-            if (file != null && !file.isEmpty()) {
-                File savedFile = fileUtil.saveFile(file);
-                teacherDTO.setTeacherProfileImg(savedFile.getName());
-            }
-        } catch (Exception e) {
-            log.info(e.getMessage());
-        }
-        adminService.insertTeacher(teacherDTO);
-        session.removeAttribute("registTeacherId");
-        session.removeAttribute("registTeacherName");
-        return "redirect:/admin/teacher/list";
-    }
-
-    @GetMapping("/teacher/modify")
-    public String teacherModify(@RequestParam(name = "teacherId", defaultValue = "") String teacherId,
-                                Model model,
-                                RedirectAttributes redirectAttributes,
-                                HttpSession session) {
-        if(teacherId.isBlank()||!adminService.existsByTeacherId(teacherId)) {
-            redirectAttributes.addFlashAttribute("errorMessage", "존재하지 않는 선생님입니다.");
-            return "redirect:/admin/teacher/list";
-        }
-
-        TeacherDTO teacherDTO = teacherService.selectTeacher(teacherId);
-        session.setAttribute("modifyTeacherId", teacherDTO.getTeacherId());
-        model.addAttribute("teacherDTO", teacherDTO);
-        setBreadcrumb(model,
-                Map.of("선생님 목록", "/admin/teacher/list"),
-                Map.of("선생님 등록", "")
-        );
-        return "admin/teacher/modify";
-    }
-
-    @PostMapping("/teacher/modify")
-    public String teacherModifyPost(@RequestParam(name = "file1") MultipartFile file,
-                                    @Valid @ModelAttribute TeacherDTO teacherDTO,
-                                    BindingResult bindingResult,
-                                    HttpSession session,
-                                    RedirectAttributes redirectAttributes) {
-        String sessionTeacherId = (String)session.getAttribute("modifyTeacherId");
-
-        if (bindingResult.hasErrors()) {
-            List<String> errorMessages = bindingResult.getFieldErrors().stream()
-                    .map(FieldError::getDefaultMessage)
-                    .collect(Collectors.toList());
-            redirectAttributes.addFlashAttribute("errorMessages", errorMessages);
-            redirectAttributes.addFlashAttribute("teacherDTO", teacherDTO);
-            return "redirect:/admin/teacher/modify?teacherId="+sessionTeacherId;
-        }
-
-        try {
-            if (file != null && !file.isEmpty()) {
-                File savedFile = fileUtil.saveFile(file);
-                if (!teacherDTO.getTeacherProfileImg().isBlank()) {
-                    fileUtil.deleteOnlyFile(teacherDTO.getTeacherProfileImg());
-                }
-                teacherDTO.setTeacherProfileImg(savedFile.getName());
-            }
-        } catch (Exception e) {
-            log.info(e.getMessage());
-        }
-        teacherDTO.setTeacherId(sessionTeacherId);
-        adminService.modifyTeacherProfile(teacherDTO);
-        session.removeAttribute("modifyTeacherId");
-        return "redirect:/admin/teacher/list";
-    }
-
-    @GetMapping("/teacher/search")
-    public String teacherSearchPopup(@ModelAttribute MemberPageDTO memberPageDTO, Model model, HttpServletRequest req) {
-        memberPageDTO.setSearch_member_grade("13");
-        memberPageDTO.setPage_size(5);
-        String baseUrl = req.getRequestURI();
-        memberPageDTO.setLinkUrl(NewPagingUtil.buildLinkUrl(baseUrl, memberPageDTO));
-        int total_count = adminService.getAllTeachersCount(memberPageDTO);
-        memberPageDTO.setTotal_count(total_count);
-        String paging = NewPagingUtil.pagingArea(memberPageDTO);
-        List<MemberDTO> memberDTOs = adminService.getAllTeachers(memberPageDTO);
-        model.addAttribute("teachers", memberDTOs);
-        model.addAttribute("searchDTO", memberPageDTO);
-        model.addAttribute("paging", paging);
-        return "/admin/teacher/searchPopup";
-    }
-
-    @PostMapping("/teacher/delete")
-    @ResponseBody
-    public Map<String, Object> teacherDelete(@RequestParam("teacherId") String teacherId) {
-        Map<String, Object> result = new HashMap<>();
-        int rtnResult = adminService.deleteTeacher(teacherId);
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "선생님 삭제되었습니다.");
-            result.put("redirect", "/admin/teacher/list");
-        } else {
-            result.put("success", false);
-            result.put("message", "삭제에 실패했습니다.");
-        }
-        return result;
-    }
-
-    @PostMapping("/teacher/restore")
-    @ResponseBody
-    public Map<String, Object> teacherRestore(@RequestParam("teacherId") String teacherId) {
-        Map<String, Object> result = new HashMap<>();
-        int rtnResult = adminService.restoreTeacher(teacherId);
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "선생님 정보가 복구되었습니다.");
-            result.put("redirect", "/admin/teacher/list");
-        } else {
-            result.put("success", false);
-            result.put("message", "삭제에 실패했습니다.");
-        }
-        return result;
-    }
-
-    @GetMapping("/lecture/list")
-    public void lectureList(@ModelAttribute LecturePageDTO lecturePageDTO, Model model) {
-        List<LectureDTO> lectureDTOs = adminService.selectLectureList(lecturePageDTO);
-        model.addAttribute("lectures", lectureDTOs);
-        setBreadcrumb(model, Map.of("강좌 목록", ""));
-    }
-
-    @GetMapping("/lecture/regist")
-    public void lectureRegist(Model model) {
-        setBreadcrumb(model,
-                Map.of("강좌 목록", "/admin/lecture/list"),
-                Map.of("강좌 등록", "")
-        );
-    }
-
-    @PostMapping("/lecture/regist")
-    public String lectureRegistPOST(@RequestParam(name = "file1") MultipartFile file,
-                                  @Valid @ModelAttribute LectureDTO lectureDTO,
-                                  BindingResult bindingResult,
-                                  RedirectAttributes redirectAttributes) {
-
-        if (bindingResult.hasErrors()) {
-            List<String> errorMessages = bindingResult.getFieldErrors().stream()
-                    .map(FieldError::getDefaultMessage)
-                    .collect(Collectors.toList());
-            redirectAttributes.addFlashAttribute("errorMessages", errorMessages);
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            log.error(errorMessages);
-            return "redirect:/admin/lecture/regist";
-        }
-
-        if (lectureDTO.getLectureAmount() < 0 || lectureDTO.getLectureAmount() > 5000000) {
-            redirectAttributes.addFlashAttribute("errorMessage", "강좌 가격을 최대 5,000,000원 이하로 입력하세요.");
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            return "redirect:/admin/lecture/regist";
-        }
-        log.info("lectureDTO.getLectureTeacherId() : {}", lectureDTO.getLectureTeacherId());
-        log.info("lectureDTO.getLectureTeacherId().isBlank() : {}", lectureDTO.getLectureTeacherId().isBlank());
-        log.info("!adminService.existsByTeacherId(lectureDTO.getLectureTeacherId()) : {}", !adminService.existsByTeacherId(lectureDTO.getLectureTeacherId()));
-        if (lectureDTO.getLectureTeacherId().isBlank() || !adminService.existsByTeacherId(lectureDTO.getLectureTeacherId())) {
-            redirectAttributes.addFlashAttribute("errorMessage", "선생님 정보가 올바르지 않습니다.");
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            return "redirect:/admin/lecture/regist";
-        }
-
-        if (file != null && !file.isEmpty() && file.getSize() > MAX_FILE_SIZE_10) {
-            redirectAttributes.addFlashAttribute("errorMessage", "파일은 10MB 이하만 업로드할 수 있습니다.");
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            log.error("파일 크기 제한");
-            return "redirect:/admin/lecture/regist";
-        }
-
-        try {
-            if (file != null && !file.isEmpty()) {
-                File savedFile = fileUtil.saveFile(file);
-                lectureDTO.setLectureThumbnailImg(savedFile.getName());
-            }
-        } catch (Exception e) {
-            log.info(e.getMessage());
-        }
-        adminService.insertLecture(lectureDTO);
-        return "redirect:/admin/lecture/list";
-    }
-
-
-    @GetMapping("/lecture/modify")
-    public void lectureModify(@RequestParam("lectureIdx") int lectureIdx, Model model) {
-        LectureDTO lectureDTO = adminService.selectLecture(lectureIdx);
-        model.addAttribute("lectureDTO", lectureDTO);
-        setBreadcrumb(model,
-                Map.of("강좌 목록", "/admin/lecture/list"),
-                Map.of("강좌 수정", "")
-        );
-    }
-
-    @PostMapping("/lecture/modify")
-    public String lectureModifyPOST(@RequestParam(name = "file1") MultipartFile file,
-                                  @RequestParam("lectureIdx") int lectureIdx,
-                                  @Valid @ModelAttribute LectureDTO lectureDTO,
-                                  BindingResult bindingResult,
-                                  RedirectAttributes redirectAttributes) {
-
-        if (bindingResult.hasErrors()) {
-            List<String> errorMessages = bindingResult.getFieldErrors().stream()
-                    .map(FieldError::getDefaultMessage)
-                    .collect(Collectors.toList());
-            redirectAttributes.addFlashAttribute("errorMessages", errorMessages);
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            log.error(errorMessages);
-            return "redirect:/admin/lecture/modify?lectureIdx="+lectureIdx;
-        }
-
-        if (lectureDTO.getLectureAmount() < 0 || lectureDTO.getLectureAmount() > 5000000) {
-            redirectAttributes.addFlashAttribute("errorMessage", "강좌 가격을 최대 5,000,000원 이하로 입력하세요.");
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            return "redirect:/admin/lecture/regist";
-        }
-
-        if (lectureDTO.getLectureTeacherId().isBlank() || !adminService.existsByTeacherId(lectureDTO.getLectureTeacherId())) {
-            redirectAttributes.addFlashAttribute("errorMessage", "선생님 정보가 올바르지 않습니다.");
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            return "redirect:/admin/lecture/modify?lectureIdx="+lectureIdx;
-        }
-
-        if (file != null && !file.isEmpty() && file.getSize() > MAX_FILE_SIZE_10) {
-            redirectAttributes.addFlashAttribute("errorMessage", "파일은 10MB 이하만 업로드할 수 있습니다.");
-            redirectAttributes.addFlashAttribute("lectureDTO", lectureDTO);
-            log.error("파일 크기 제한");
-            return "redirect:/admin/lecture/modify?lectureIdx="+lectureIdx;
-        }
-
-        try {
-            if (file != null && !file.isEmpty()) {
-                File savedFile = fileUtil.saveFile(file);
-                lectureDTO.setLectureThumbnailImg(savedFile.getName());
-            }
-        } catch (Exception e) {
-            log.info(e.getMessage());
-        }
-        adminService.updateLecture(lectureDTO);
-        return "redirect:/admin/lecture/list";
-    }
-
-    @PostMapping("/lecture/delete")
-    @ResponseBody
-    public Map<String, Object> lectureDelete(@RequestParam("lectureIdx") int lectureIdx) {
-        Map<String, Object> result = new HashMap<>();
-        int rtnResult = adminService.deleteLecture(lectureIdx);
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "강좌가 삭제되었습니다.");
-            result.put("redirect", "/admin/lecture/list");
-        } else {
-            result.put("success", false);
-            result.put("message", "삭제에 실패했습니다.");
-        }
-        return result;
-    }
-
-    @PostMapping("/lecture/restore")
-    @ResponseBody
-    public Map<String, Object> lectureRestore(@RequestParam("lectureIdx") int lectureIdx) {
-        Map<String, Object> result = new HashMap<>();
-        int rtnResult = adminService.restoreLecture(lectureIdx);
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "강좌가 복구되었습니다.");
-            result.put("redirect", "/admin/lecture/list");
-        } else {
-            result.put("success", false);
-            result.put("message", "삭제에 실패했습니다.");
-        }
-        return result;
-    }
-
-    @GetMapping("/lecture/search")
-    public String lectureSearchPopup(@ModelAttribute LecturePageDTO lecturePageDTO, HttpServletRequest req, Model model) {
-        String baseUrl = req.getRequestURI();
-        lecturePageDTO.setLinkUrl(NewPagingUtil.buildLinkUrl(baseUrl, lecturePageDTO));
-        int total_count = adminService.selectLectureCount(lecturePageDTO);
-        lecturePageDTO.setTotal_count(total_count);
-        String paging = NewPagingUtil.pagingArea(lecturePageDTO);
-        List<LectureDTO> lectureDTOs = adminService.selectLectureList(lecturePageDTO);
-        model.addAttribute("lectures", lectureDTOs);
-        model.addAttribute("searchDTO", lecturePageDTO);
-        model.addAttribute("paging", paging);
-        return "/admin/lecture/searchPopup";
-    }
-
-    @GetMapping("/chapter/list")
-    public void chapterList(@ModelAttribute ChapterPageDTO chapterPageDTO, Model model, HttpServletRequest req) {
-        String baseUrl = req.getRequestURI();
-        chapterPageDTO.setLinkUrl(NewPagingUtil.buildLinkUrl(baseUrl, chapterPageDTO));
-        int totalCount = adminService.selectChapterCount(chapterPageDTO);
-        String paging = NewPagingUtil.pagingArea(chapterPageDTO);
-        List<ChapterDTO> chapters = adminService.selectChapterList(chapterPageDTO);
-        model.addAttribute("chapters", chapters);
-        model.addAttribute("totalCount", totalCount);
-        model.addAttribute("searchDTO", chapterPageDTO);
-        model.addAttribute("paging", paging);
-        model.addAttribute("lectureIdx", chapterPageDTO.getLectureIdx());
-        setBreadcrumb(model,
-                Map.of("강좌 목록", "/admin/lecture/list"),
-                Map.of("강의 목록", "")
-        );
-    }
-
-    @GetMapping("/chapter/regist")
-    public void chapterRegist(@RequestParam(name = "lectureIdx", defaultValue = "0") int lectureIdx, Model model) {
-        LectureDTO lectureDTO = adminService.selectLecture(lectureIdx);
-        model.addAttribute("lectureIdx", lectureIdx);
-        model.addAttribute("lectureDTO", lectureDTO);
-        setBreadcrumb(model,
-                Map.of("강좌 목록", "/admin/lecture/list"),
-                Map.of("강의 등록", "")
-        );
-    }
-
-    @PostMapping("/chapter/regist")
-    @ResponseBody
-    public ResponseEntity<?> chapterRegistPOST(@RequestParam(name = "file1", required = true) MultipartFile file,
-                                               @Valid @ModelAttribute ChapterDTO chapterDTO,
-                                               BindingResult bindingResult) {
-        if(bindingResult.hasErrors()) {
-            String errorMessage = bindingResult.getFieldError().getDefaultMessage();
-            log.error(errorMessage);
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(Map.of("message", errorMessage));
-        }
-
-        if(file==null || file.isEmpty())
-            return ResponseEntity
-                .status(HttpStatus.BAD_REQUEST)
-                .body(Map.of("message", "강의 영상은 필수입니다."));
-
-        if(adminService.existsByLectureId(chapterDTO.getChapterLectureIdx()))
-            return ResponseEntity
-                    .status(HttpStatus.BAD_REQUEST)
-                    .body(Map.of("message", "존재하지 않는 강좌 번호입니다."));
-
-        try {
-            if (file != null && !file.isEmpty()) {
-                if (file.getSize()>MAX_FILE_SIZE_100) {
-                    return ResponseEntity
-                            .status(HttpStatus.PAYLOAD_TOO_LARGE)
-                            .body(Map.of("message", "파일 크기가 100MB를 초과했습니다."));
-                }
-                File savedFile = fileUtil.saveFile(file);
-                chapterDTO.setChapterVideo(savedFile.getName());
-                chapterDTO.setChapterRuntime(VideoUtil.getVideoDuration(savedFile));
-                adminService.insertChapter(chapterDTO);
-                return ResponseEntity.ok(Map.of("message", "강의 등록 완료"));
-            }
-        } catch (Exception e) {
-            log.error(e.getMessage());
-        }
-        return ResponseEntity
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(Map.of("message", "강의 등록 중 오류가 발생했습니다."));
-    }
-
-    @GetMapping("/chapter/modify")
-    public void chapterModify(@RequestParam("chapterIdx") int chapterIdx, Model model) {
-        ChapterDTO chapterDTO = adminService.selectChapter(chapterIdx);
-        model.addAttribute("chapterDTO", chapterDTO);
-        setBreadcrumb(model,
-                Map.of("강의 목록", "/admin/chapter/list"),
-                Map.of("강의 수정", "")
-        );
-    }
-
-    @PostMapping("/chapter/modify")
-    @ResponseBody
-    public ResponseEntity<?> chapterModifyPOST(@RequestParam(name = "file1") MultipartFile file,
-                                               @Valid @ModelAttribute ChapterDTO chapterDTO,
-                                               BindingResult bindingResult) {
-        if(bindingResult.hasErrors()) {
-            String errorMessage = bindingResult.getFieldError().getDefaultMessage();
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorMessage);
-        }
-
-        if(adminService.existsByLectureId(chapterDTO.getChapterLectureIdx()))
-            return ResponseEntity
-                    .status(HttpStatus.BAD_REQUEST)
-                    .body(Map.of("message", "강좌 번호가 존재하지 않습니다."));
-
-        try {
-            if (file != null && !file.isEmpty()) {
-                if (file.getSize() > MAX_FILE_SIZE_100) {
-                    return ResponseEntity
-                            .status(HttpStatus.PAYLOAD_TOO_LARGE)
-                            .body(Map.of("message", "파일 크기가 100MB를 초과했습니다."));
-                }
-                File savedFile = fileUtil.saveFile(file);
-                chapterDTO.setChapterVideo(savedFile.getName());
-                chapterDTO.setChapterRuntime(VideoUtil.getVideoDuration(savedFile));
-            }
-        } catch (Exception e) {
-            log.error(e.getMessage());
-        }
-        adminService.updateChapter(chapterDTO);
-        return ResponseEntity.ok(Map.of("message", "강의 수정 성공"));
-    }
-
-    @PostMapping("/chapter/delete")
-    @ResponseBody
-    public Map<String, Object> chapterDelete(@RequestParam("chapterIdx") int chapterIdx) {
-        Map<String, Object> result = new HashMap<>();
-        int rtnResult = adminService.deleteChapter(chapterIdx);
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "강의가 삭제되었습니다.");
-            result.put("redirect", "/admin/chapter/list");
-        } else {
-            result.put("success", false);
-            result.put("message", "삭제에 실패했습니다.");
-        }
-        return result;
-    }
-
-    @PostMapping("/chapter/restore")
-    @ResponseBody
-    public Map<String, Object> chapterRestore(@RequestParam("chapterIdx") int chapterIdx) {
-        Map<String, Object> result = new HashMap<>();
-        int rtnResult = adminService.restoreChapter(chapterIdx);
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "강의가 복구되었습니다.");
-            result.put("redirect", "/admin/chapter/list");
-        } else {
-            result.put("success", false);
-            result.put("message", "삭제에 실패했습니다.");
-        }
-        return result;
-    }
-
-    @GetMapping("/sales/info")
-    public String salesInfo(@ModelAttribute SearchDTO searchDTO,
-                            @ModelAttribute PageRequestDTO pageRequestDTO,
-                            Model model) {
-
-        // 월별/강좌별 매출은 기존대로
-        List<Map<String, Object>> monthlySales = adminService.getMonthlySales();
-        List<Map<String, Object>> lectureSales = adminService.getLectureSales();
-
-        // 개별 거래 내역 (조건 기반)
-        List<OrderDTO> detailList = adminService.getSalesDetailList(searchDTO, pageRequestDTO);
-        int count = adminService.getSalesDetailCount(searchDTO);
-
-        PageResponseDTO<OrderDTO> pageResponseDTO = PageResponseDTO.<OrderDTO>withAll()
-                .dtoList(detailList)
-                .totalCount(count)
-                .pageRequestDTO(pageRequestDTO)
-                .build();
-
-        model.addAttribute("monthlySales", monthlySales);
-        model.addAttribute("lectureSales", lectureSales);
-        model.addAttribute("searchDTO", searchDTO);
-        model.addAttribute("responseDTO", pageResponseDTO);
-
-        setBreadcrumb(model, Map.of("매출 정보", ""));
-        return "/admin/sales/dashboard";
-    }
-
-
-    @GetMapping("/sales/monthly")
-    @ResponseBody
-    public List<Map<String, Object>> getMonthlySales(
-            @RequestParam("timeType") String timeType,
-            @RequestParam(value = "startDate", required = false) String startDate,
-            @RequestParam(value = "endDate", required = false) String endDate
-    ) {
-        return adminService.getMonthlySales(timeType, startDate, endDate);
-    }
-
-    @GetMapping("/sales/lecture")
-    @ResponseBody
-    public List<Map<String, Object>> getLectureSales(
-            @RequestParam("timeType") String timeType,
-            @RequestParam("startDate") String startDate,
-            @RequestParam("endDate") String endDate) {
-        return adminService.getLectureSales(timeType, startDate, endDate);
-    }
-
-    @GetMapping("/sales/detail")
-    @ResponseBody
-    public PageResponseDTO<OrderDTO> salesDetailList(
-            @RequestParam(value = "searchWord", required = false) String searchWord,
-            @RequestParam(value = "searchType", required = false) String searchType,
-            @RequestParam(value = "startDate", required = false) String startDate,
-            @RequestParam(value = "endDate", required = false) String endDate,
-            @RequestParam(value = "pageNo", defaultValue = "1") int pageNo,
-            @RequestParam(value = "pageSize", defaultValue = "10") int pageSize,
-            @RequestParam(value = "sortColumn", required = false) String sortColumn,
-            @RequestParam(value = "sortOrder", required = false) String sortOrder
-    ) {
-        SearchDTO searchDTO = new SearchDTO();
-        searchDTO.setSearchWord(searchWord);
-        searchDTO.setSearchType(searchType);
-        searchDTO.setSortColumn(sortColumn);
-        searchDTO.setSortOrder(sortOrder);
-        searchDTO.setStartDate(startDate);
-        searchDTO.setEndDate(endDate);
-
-        PageRequestDTO pageRequestDTO = PageRequestDTO.builder()
-                .pageNo(pageNo)
-                .pageSize(pageSize)
-                .build();
-
-        List<OrderDTO> detailList = adminService.getSalesDetailList(searchDTO, pageRequestDTO);
-        int count = adminService.getSalesDetailCount(searchDTO);
-
-        return PageResponseDTO.<OrderDTO>withAll()
-                .dtoList(detailList)
-                .totalCount(count)
-                .pageRequestDTO(pageRequestDTO)
-                .build();
-    }
-
-    @GetMapping("/sales/export")
-    public void exportSalesToExcel(
-            @RequestParam(name = "searchType", required = false) String searchType,
-            @RequestParam(name = "searchWord", required = false) String searchWord,
-            @RequestParam(name = "startDate", required = false) String startDate,
-            @RequestParam(name = "endDate", required = false) String endDate,
-            HttpServletResponse response
-    ) throws IOException {
-
-        Map<String, Object> param = new HashMap<>();
-        param.put("searchType", searchType);
-        param.put("searchWord", searchWord);
-        param.put("startDate", startDate);
-        param.put("endDate", endDate);
-
-        List<OrderDTO> list = adminService.getSalesListForExport(param);
-
-        Workbook workbook = new XSSFWorkbook();
-        Sheet sheet = workbook.createSheet("매출 내역");
-
-        Row header = sheet.createRow(0);
-        header.createCell(0).setCellValue("주문번호");
-        header.createCell(1).setCellValue("회원 ID");
-        header.createCell(2).setCellValue("강좌명");
-        header.createCell(3).setCellValue("금액");
-        header.createCell(4).setCellValue("주문일");
-
-        int rowNum = 1;
-        for (OrderDTO row : list) {
-            Row r = sheet.createRow(rowNum++);
-            r.createCell(0).setCellValue(row.getOrderIdx());
-            r.createCell(1).setCellValue(row.getOrderMemberId());
-            r.createCell(2).setCellValue(row.getLectureTitle());
-            r.createCell(3).setCellValue(row.getOrderAmount());
-            r.createCell(4).setCellValue(row.getOrderCreatedAt().toString());
-        }
-
-        response.setContentType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet");
-        response.setHeader("Content-Disposition", "attachment; filename=\"sales.xlsx\"");
-
-        workbook.write(response.getOutputStream());
-        workbook.close();
-    }
-
-    // Board manage
-    @GetMapping("/board/manage")
-    public String boardReportList(@ModelAttribute PostPageDTO postPageDTO,
-                                  Model model,
-                                  HttpServletRequest req) {
-        postPageDTO.setPostCategory(BoardCategory.freeboard.name().toUpperCase());
-        String baseUrl = req.getRequestURI();
-        postPageDTO.normalizeSearchType();
-        postPageDTO.setLinkUrl(PagingUtil.buildLinkUrl(baseUrl, postPageDTO));
-        int totalCount = boardService.getPostCount(postPageDTO);
-        postPageDTO.setTotal_count(totalCount);
-        String paging = NewPagingUtil.pagingArea(postPageDTO);
-        List<PostDTO> postDTOs = adminService.selectReportedPosts(postPageDTO);
-        model.addAttribute("postTotalCount", totalCount);
-        model.addAttribute("posts", postDTOs);
-        model.addAttribute("search", postPageDTO);
-        model.addAttribute("paging", paging);
-        setBreadcrumb(model, Map.of("자유게시판 관리", ""));
-        return "/admin/board/list";
-    }
-
-    @GetMapping("/board/manage/view")
-    @ResponseBody
-    public ResponseEntity<PostDTO> boardManageView(@RequestParam("idx") int idx) {
-        PostDTO postDTO = adminService.selectPostByIdx(idx);
-        return ResponseEntity.ok(postDTO);
-    }
-
-    @PostMapping("/board/report/state")
-    public Map<String, Object> boardReportState() {
-        Map<String, Object> result = new HashMap<>();
-        int rtnResult = 0;
-        //adminService.updateReport();
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "신고가 접수되었습니다.");
-        } else {
-            result.put("success", false);
-            result.put("message", "신고 접수에 실패했습니다.");
-        }
-
-        return result;
-    }
-
-    @PostMapping("/board/delete")
-    @ResponseBody
-    public Map<String, Object> boardReportDelete(@RequestParam("postIdx") int postIdx) {
-        Map<String, Object> result = new HashMap<>();
-
-        PostDTO postDTO = adminService.selectPostByIdx(postIdx);
-
-        if (postDTO.getPostState() == 3) {
-            result.put("success", false);
-            result.put("message", "이미 삭제된 게시글입니다.");
-            return result;
-        }
-
-        int rtnResult = adminService.deletePostByAdmin(postIdx);
-        if (rtnResult > 0) {
-            result.put("success", true);
-            result.put("message", "게시글이 삭제되었습니다.");
-        } else {
-            result.put("success", false);
-            result.put("message", "게시글 삭제에 실패했습니다.");
-        }
-        return result;
-    }
-
-    @GetMapping("/qna/list")
-    public String qna(@ModelAttribute SearchDTO searchDTO,
-                      @ModelAttribute PageRequestDTO pageRequestDTO,
-                      Model model) {
-
-        int totalCount = qnaService.notAnsweredQnaTotalCount(searchDTO);
-        PageResponseDTO<QnaDTO> pageResponseDTO = PageUtil.buildAndCorrectPageResponse(
-                pageRequestDTO,
-                totalCount,
-                () -> qnaService.notAnsweredQna(searchDTO, pageRequestDTO)
-        );
-
-        model.addAttribute("responseDTO", pageResponseDTO);
-        model.addAttribute("notAnsweredQnaList", pageResponseDTO.getDtoList());
-        model.addAttribute("searchDTO", searchDTO);
-        setBreadcrumb(model, Map.of("미답변 문의 관리", ""));
-
-        return "/admin/qna/list";
-    }
-
-    // 브레드크럼
-    private void setBreadcrumb(Model model, Map<String, String>... pagePairs) {
-        LinkedHashMap<String, String> pages = new LinkedHashMap<>();
-        for (Map<String, String> page : pagePairs) {
-            pages.putAll(page);
-        }
-        BreadcrumbUtil.addBreadcrumb(model, pages, ROOT_BREADCRUMB);
-    }
-
-}
-
