+++ src/main/java/net/spb/spb/controller/admin/AdminMemberController.java
@@ -1,0 +1,95 @@
+package net.spb.spb.controller.admin;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import net.spb.spb.dto.ChapterDTO;
+import net.spb.spb.dto.LectureDTO;
+import net.spb.spb.dto.OrderDTO;
+import net.spb.spb.dto.TeacherDTO;
+import net.spb.spb.dto.member.MemberDTO;
+import net.spb.spb.dto.pagingsearch.*;
+import net.spb.spb.dto.post.PostDTO;
+import net.spb.spb.dto.post.PostReportDTO;
+import net.spb.spb.dto.qna.QnaDTO;
+import net.spb.spb.service.AdminService;
+import net.spb.spb.service.ReportService;
+import net.spb.spb.service.board.BoardServiceIf;
+import net.spb.spb.service.member.MemberServiceIf;
+import net.spb.spb.service.qna.QnaService;
+import net.spb.spb.service.teacher.TeacherServiceIf;
+import net.spb.spb.util.*;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+@Controller
+@Log4j2
+@RequiredArgsConstructor
+@RequestMapping("/admin")
+public class AdminMemberController extends AdminBaseController {
+
+    private final MemberServiceIf memberService;
+    private final AdminService adminService;
+
+    @GetMapping("/member/list")
+    public void memberList(@ModelAttribute MemberPageDTO memberPageDTO, Model model, HttpServletRequest req) {
+        String baseUrl = req.getRequestURI();
+        memberPageDTO.setLinkUrl(PagingUtil.buildLinkUrl(baseUrl, memberPageDTO));
+        if (memberPageDTO.getSearch_member_grade() != null && memberPageDTO.getSearch_member_grade().trim().equals("")) {
+            memberPageDTO.setSearch_member_grade(null);
+        }
+        int total_count = memberService.getMemberCount(memberPageDTO);
+        memberPageDTO.setTotal_count(total_count);
+        String paging = NewPagingUtil.pagingArea(memberPageDTO);
+        List<MemberDTO> memberDTOs = memberService.getMembers(memberPageDTO);
+        model.addAttribute("searchDTO", memberPageDTO);
+        model.addAttribute("list", memberDTOs);
+        model.addAttribute("memberTotalCount", total_count);
+        model.addAttribute("paging", paging);
+        setBreadcrumb(model, Map.of("회원 목록", ""));
+    }
+
+    @PostMapping("/member/update")
+    public String memberModify(@ModelAttribute MemberDTO memberDTO) {
+        memberService.updateMemberByAdmin(memberDTO);
+        return "/admin/member/update-success";
+    }
+
+    @PostMapping("/member/state")
+    public String memberModifyState(@ModelAttribute MemberPageDTO memberPageDTO, @ModelAttribute MemberDTO memberDTO) {
+        memberService.updateMemberState(memberDTO);
+        String query = memberPageDTO.toQueryString();
+        return "redirect:/admin/member/list?" + query;
+    }
+
+    @GetMapping("/member/view")
+    public String memberView(@RequestParam("memberId") String memberId, Model model) {
+        MemberDTO memberDTO = memberService.getMemberById(memberId);
+        model.addAttribute("memberDTO", memberDTO);
+        return "/admin/member/memberPopup";
+    }
+
+}
+
