+++ src/main/webapp/WEB-INF/views/admin/teacher/requestList.jsp
@@ -1,3 +1,10 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: sinjihye
+  Date: 2025. 5. 15.
+  Time: 23:15
+  To change this template use File | Settings | File Templates.
+--%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form" %>
 <%@ taglib prefix="fn" uri="http://java.sun.com/jsp/jstl/functions" %>
@@ -16,69 +23,51 @@
     <div class="container my-5">
         <%@ include file="../../common/breadcrumb.jsp" %>
     </div>
-
-    <c:if test="${not empty teacher2}">
-    <div class="container my-5">
-        <div class="alert alert-warning alert-dismissible fade show" role="alert">
-            <b>선생님 등록 요청 ${fn:length(teacher2)}건</b>
-            <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
-        </div>
+    <div class="container mt-5 mb-2 d-flex justify-content-between">
+        <h2 class="h4 fw-bold">선생님 요청 목록</h2>
+        <select class="form-select form-select-sm w-auto" name="page_size" id="selectPageSize">
+            <option value="5" ${search.page_size == 5 ? "selected":""}>5개씩</option>
+            <option value="10" ${search.page_size == 10 ? "selected":""}>10개씩</option>
+            <option value="15" ${search.page_size == 15 ? "selected":""}>15개씩</option>
+            <option value="20" ${search.page_size == 20 ? "selected":""}>20개씩</option>
+            <option value="30" ${search.page_size == 30 ? "selected":""}>30개씩</option>
+        </select>
     </div>
+    <c:if test="${not empty teacher2}">
+        <div class="container my-4">
+            <div class="alert alert-warning alert-dismissible fade show" role="alert">
+                <b>선생님 등록 요청 ${totalCount}건 있습니다.</b>
+                <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
+            </div>
+        </div>
     </c:if>
 
-    <div class="container mt-5">
-        <h2 class="h4 fw-bold">선생님 목록</h2>
-    </div>
-
-    <div class="container my-2" style="max-height: 100vh;">
-        <div class="text-muted my-2">${totalCount}명의 선생님이 등록되어 있습니다.</div>
+    <div class="container my-2 pb-5" style="min-height: 100vh;">
         <table class="table table-hover text-center align-middle">
             <thead class="table-light">
-                <tr>
-                    <th scope="col">No</th>
-                    <th scope="col">과목</th>
-                    <th scope="col">이름</th>
-                    <th scope="col">생년월일</th>
-                    <th scope="col">회원 가입일</th>
-                    <th scope="col">선생님 페이지</th>
-                    <th scope="col">상태</th>
-                    <th scope="col">정보 수정</th>
-                </tr>
+            <tr id="teacherRow-${memberId}">
+                <th scope="col">No</th>
+                <th scope="col">아이디</th>
+                <th scope="col">이름</th>
+                <th scope="col">생년월일</th>
+                <th scope="col">회원 가입일</th>
+                <th scope="col">정보 등록</th>
+            </tr>
             </thead>
             <tbody>
             <c:choose>
-                <c:when test="${not empty teacher1}">
-                    <c:forEach items="${teacher1}" var="member" varStatus="status">
-                        <tr class="clickable-row" data-href="/teacher/personal?teacherId=${member.memberId}">
+                <c:when test="${not empty teacher2}">
+                    <c:forEach items="${teacher2}" var="member" varStatus="status">
+                        <tr>
                             <td>${status.count} </td>
-                            <td class="text-truncate" style="max-width: 60px;">${member.teacherSubject}</td>
-                            <td>${member.memberName}(${member.memberId})</td>
+                            <td>${member.memberId}</td>
+                            <td>${member.memberName}</td>
                             <td>
                                 <c:set var="rawDate" value="${member.memberBirth}" />
                                     ${fn:substring(rawDate, 0, 4)}-${fn:substring(rawDate, 4, 6)}-${fn:substring(rawDate, 6, 8)}
                             </td>
                             <td>${member.memberCreatedAt}</td>
-                            <td>
-                                <a href="/teacher/personal?teacherId=${member.memberId}" class="btnModifyTeacher btn btn-sm btn-secondary"><i class="bi bi-person-badge"></i> 이동</a>
-                            <td>
-                                <c:choose>
-                                    <c:when test="${member.teacherState == 1}"><span class="badge bg-success">정상</span></c:when>
-                                    <c:otherwise><span class="badge bg-secondary">삭제됨</span></c:otherwise>
-                                </c:choose>
-                            </td>
-                            <td>
-                                <c:choose>
-                                    <c:when test="${member.teacherState == 1}">
-                                        <button type="button" data-member-id="${member.memberId}" class="btnModifyTeacher btn btn-sm btn-warning"><i class="bi bi-pencil"></i> 수정</button>
-                                        <button type="button" class="btn btn-sm btn-danger btnDelete" data-teacher-id="${member.memberId}"><i class="bi bi-trash"></i> 삭제</button>
-                                    </c:when>
-                                    <c:otherwise>
-                                        <button type="button" class="btn btn-sm btn-danger btnRestoreTeacher"
-                                                data-teacher-id="${member.memberId}">
-                                            <i class="bi bi-arrow-counterclockwise"></i> 복구
-                                        </button>
-                                    </c:otherwise>
-                                </c:choose>
+                            <td><button type="button" data-member-id="${member.memberId}" class="btnRegistTeacher btn btn-sm btn-outline-success">등록</button>
                             </td>
                         </tr>
                     </c:forEach>
@@ -97,6 +86,7 @@
     </div>
 </div>
 <jsp:include page="/WEB-INF/views/common/toast.jsp" />
+
 <script src="${pageContext.request.contextPath}/resources/js/toast.js"></script>
 <script>
     document.addEventListener('DOMContentLoaded', function () {
@@ -117,6 +107,13 @@
                 const memberId = btn.getAttribute('data-member-id');
                 window.location.href = 'modify?teacherId=' + memberId;
             });
+        });
+
+        document.getElementById("selectPageSize").addEventListener('change', function () {
+            const pageSize = this.value;
+            const url = new URL(window.location.href);
+            url.searchParams.set("page_size", pageSize);
+            window.location.href = url.toString();
         });
     });
